+++ R/officer.pptx-package.R
@@ -2,5 +2,12 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom xml2 xml_attr
+#' @importFrom xml2 xml_attr<-
+#' @importFrom xml2 xml_child
+#' @importFrom xml2 xml_find_all
+#' @importFrom xml2 xml_name
+#' @importFrom xml2 xml_remove
+#' @importFrom xml2 xml_text
 ## usethis namespace: end
 NULL